+++ src/commands/firestore-databases-get.ts
@@ -11,11 +11,6 @@ import { FirestoreOptions } from "../firestore/options";
 export const command = new Command("firestore:databases:get")
   .description("Get database in your Cloud Firestore project.")
   .option(
-    "--pretty",
-    "Pretty print. When not specified the database is printed in the " +
-      "JSON specification format."
-  )
-  .option(
     "--database <databaseId>",
     "Database ID of the firestore database from which to list configuration. (default) if none provided."
   )
@@ -27,12 +22,6 @@ export const command = new Command("firestore:databases:get")
     const databaseId = options.database || "(default)";
     const database: types.DatabaseResp = await api.getDatabase(options.project, databaseId);
 
-    if (options.pretty) {
-      logger.info(clc.bold(clc.white("Firestore Databases")));
-      api.prettyPrintDatabases([database]);
-    } else {
-      logger.info(JSON.stringify(database, undefined, 2));
-    }
-
+    logger.info(JSON.stringify(database, undefined, 2));
     return database;
   });